workers/scraper/requirements.txt
scrapling>=0.2.9
redis>=5.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
loguru>=0.7.2
schedule>=1.2.2
